+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2015-08-19
 
 - fix migration issue (https://github.com/getsentry/sentry/issues/1648)
+- allow for a simple database accessibility check (via ``SENTRY_DOCKER_DO_DB_CHECK`` env var)
 
 ## 2015-06-25